Implementación de Knuckles, Tails, y Eggman
Este merge request incorpora mejoras significativas en el juego, enfocándose tanto en el comportamiento del enemigo como en la adición de nuevos personajes jugables.

Se refinó el movimiento de Eggman (RobotnikVisual) en PantallaDeJuego.java. Inicialmente, se implementó una lógica de desatasco más agresiva para evitar que Eggman quedara atrapado en las estructuras del mapa, aumentando la duración y velocidad de su movimiento forzado (UNSTICK_DURATION a 8.0s y UNSTICK_SPEED_MULTIPLIER a 3.5x). Sin embargo, la estrategia se ajustó para que Eggman ya no colisione con los objetos del mapa. Como resultado, la lógica de detección y resolución de colisiones, así como el sistema de desatasco previamente implementado, fueron eliminados, permitiendo que Eggman se mueva libremente y atraviese cualquier estructura mientras persigue a Sonic, impulsado únicamente por su VELOCIDAD_ROBOTNIK.

Además, este merge introduce a los personajes Knuckles y Tails. Se implementaron sus respectivas clases, lo que incluye la integración de sus sprites y animaciones para que puedan ser representados y posiblemente controlados o actúen en el juego.

+++ core/src/main/java/network/LocalServer.java
@@ -41,7 +41,7 @@ public class LocalServer implements IGameServer {
     // Referencia directa al único cliente que existirá en este modo
     private LocalClient clienteLocal;
     private int proximoIdJugador = 1; // En modo local, siempre empezamos en 1
-    private static final float ROBOT_SPEED = 1.0f;
+    private static final int ROBOT_SPEED = 1;
     private static final float ROBOT_DETECTION_RANGE = 300f;
     private static final float ROBOT_ATTACK_RANGE = 10f; // Usando el valor del código original
 
@@ -131,9 +131,9 @@ public class LocalServer implements IGameServer {
         if (jugador == null) return;
 
         for (EnemigoState enemigo : enemigosActivos.values()) {
-            float dx = jugador.x - enemigo.x;
-            float dy = jugador.y - enemigo.y;
-            float distance = (float) Math.sqrt(dx * dx + dy * dy);
+            int dx = (int) (jugador.x - enemigo.x);
+            int dy = (int) (jugador.y - enemigo.y);
+            int distance = (int) Math.sqrt(dx * dx + dy * dy);
 
             // Si el robot está en medio de la animación de GOLPE, el servidor no hace NADA.
             // Simplemente espera a que el cliente le notifique que la animación ha terminado.
